+++ examples/pydy_double_pendulum.py
@@ -12,6 +12,7 @@ $ python pydy_double_pendulum.py --plot --nt 200
 import numpy as np
 
 from pyodesys import SymbolicSys
+from pyodesys.util import stack_1d_on_left
 
 
 def get_equations(m_val, g_val, l_val):
@@ -68,18 +69,20 @@ def get_equations(m_val, g_val, l_val):
 
 def main(m=1, g=9.81, l=1, q1=.1, q2=.2, u1=0, u2=0, tend=10., nt=200,
          savefig='None', plot=False, savetxt='None', solver='scipy',
-         dpi=100, kwargs=""):
+         dpi=100, kwargs="", verbose=False):
     assert nt > 1
     kwargs = dict(eval(kwargs) if kwargs else {})
     odesys = SymbolicSys(get_equations(m, g, l))
     tout = np.linspace(0, tend, nt)
     y0 = [q1, q2, u1, u2]
-    out = odesys.integrate(solver, tout, y0, **kwargs)
+    xout, yout, info = odesys.integrate(solver, tout, y0, **kwargs)
+    if verbose:
+        print(info)
     if savetxt != 'None':
-        np.savetxt(out, savetxt)
+        np.savetxt(stack_1d_on_left(xout, yout), savetxt)
     if plot:
         import matplotlib.pyplot as plt
-        plt.plot(out[:, 0], out[:, 1:])
+        odesys.plot_result(xout, yout)
         if savefig != 'None':
             plt.savefig(savefig, dpi=dpi)
         else: