+++ pyodesys/tests/test_symbolic.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 from __future__ import print_function, absolute_import, division
 
 import math
@@ -30,34 +31,33 @@ def decay_rhs(t, y, k):
     return dydt
 
 
-def _test_TransformedSys(dep_tr, indep_tr, rtol, atol):
+def _test_TransformedSys(dep_tr, indep_tr, rtol, atol, first_step, forgive=1):
     k = [7., 3, 2]
     ts = symmetricsys(dep_tr, indep_tr).from_callback(
         decay_rhs, len(k)+1, len(k))
     y0 = [1e-20]*(len(k)+1)
     y0[0] = 1
-    out = ts.integrate_cvode([1e-12, 1], y0, k, atol=1e-12, rtol=1e-12)
-    ref = out.copy()
-    ref[:, 1:] = np.array(bateman_full(y0, k+[0], ref[:, 0] - ref[0, 0],
-                                       exp=np.exp)).T
+    xout, yout, info = ts.integrate_cvode(
+        [1e-12, 1], y0, k, atol=atol, rtol=rtol, first_step=first_step)
+    ref = np.array(bateman_full(y0, k+[0], xout - xout[0], exp=np.exp)).T
     np.set_printoptions(linewidth=240)
-    assert np.allclose(out, ref, rtol=rtol, atol=atol)
+    assert np.allclose(yout, ref, rtol=rtol*forgive, atol=atol*forgive)
 
 
 def test_TransformedSys_liny_linx():
-    _test_TransformedSys(idty2, idty2, 1e-11, 1e-11)
-
-
-def test_TransformedSys_logy_linx():
-    _test_TransformedSys(logexp, idty2, 1e-10, 1e-10)
-
-
-def test_TransformedSys_liny_logx():
-    _test_TransformedSys(idty2, logexp, 1e-9, 1e-9)
+    _test_TransformedSys(idty2, idty2, 1e-11, 1e-11, 0, 10)
 
 
 def test_TransformedSys_logy_logx():
-    _test_TransformedSys(logexp, logexp, 1e-10, 1e-10)
+    _test_TransformedSys(logexp, logexp, 1e-7, 1e-7, 1e-4, 100)
+
+
+def test_TransformedSys_logy_linx():
+    _test_TransformedSys(logexp, idty2, 1e-8, 1e-8, 0, 100)
+
+
+def test_TransformedSys_liny_logx():
+    _test_TransformedSys(idty2, logexp, 1e-9, 1e-9, 0, 100)
 
 
 def test_ScaledSys():
@@ -72,11 +72,10 @@ def test_ScaledSys():
     ss = ScaledSys(l, dep_scaling=1e8)
     y0 = [0]*(len(k)+1)
     y0[0] = 1
-    out = ss.integrate_cvode([1e-12, 1], y0, atol=1e-12, rtol=1e-12)
-    ref = out.copy()
-    ref[:, 1:] = np.array(bateman_full(y0, k+[0], ref[:, 0] - ref[0, 0],
-                                       exp=np.exp)).T
-    assert np.allclose(out, ref, rtol=1e-12, atol=1e-12)
+    xout, yout, info = ss.integrate_cvode([1e-12, 1], y0,
+                                          atol=1e-12, rtol=1e-12)
+    ref = np.array(bateman_full(y0, k+[0], xout - xout[0], exp=np.exp)).T
+    assert np.allclose(yout, ref, rtol=1e-12, atol=1e-12)
 
 
 def test_ScaledSys_from_callback():
@@ -89,11 +88,9 @@ def test_ScaledSys_from_callback():
     k = [7, 3, 2]
     y0 = [0]*(len(k)+1)
     y0[0] = 1
-    out = odesys.integrate_scipy([1e-12, 1], y0, k)
-    ref = out.copy()
-    ref[:, 1:] = np.array(bateman_full(y0, k+[0], ref[:, 0] - ref[0, 0],
-                                       exp=np.exp)).T
-    assert np.allclose(out, ref, rtol=3e-11, atol=3e-11)
+    xout, yout, info = odesys.integrate_scipy([1e-12, 1], y0, k)
+    ref = np.array(bateman_full(y0, k+[0], xout - xout[0], exp=np.exp)).T
+    assert np.allclose(yout, ref, rtol=3e-11, atol=3e-11)
 
 
 def timeit(callback, *args, **kwargs):
@@ -107,11 +104,11 @@ def test_exp(method):
     x = sp.Symbol('x')
     symsys = SymbolicSys([(x, sp.exp(x))])
     tout = [0, 1e-9, 1e-7, 1e-5, 1e-3, 0.1]
-    out = symsys.integrate_odeint(tout, [1], method=method,
-                                  atol=1e-12, rtol=1e-12)
+    xout, yout, info = symsys.integrate_odeint(tout, [1], method=method,
+                                               atol=1e-12, rtol=1e-12)
     e = math.e
     ref = -math.log(1/e - 0.1)
-    assert abs(out[-1, 1] - ref) < 4e-8
+    assert abs(yout[-1, 0] - ref) < 4e-8
 
 
 # @pytest.mark.xfail
@@ -120,10 +117,10 @@ def _test_mpmath():  # too slow
     symsys = SymbolicSys([(x, sp.exp(x))])
     tout = [0, 1e-9, 1e-7, 1e-5, 1e-3, 0.1]
     # import pudb; pudb.set_trace()
-    out = symsys.integrate_mpmath(tout, [1])
+    xout, yout, info = symsys.integrate_mpmath(tout, [1])
     e = math.e
     ref = -math.log(1/e - 0.1)
-    assert abs(out[-1, 1] - ref) < 4e-8
+    assert abs(yout[-1, 0] - ref) < 4e-8
 
 
 def decay_dydt_factory(k):
@@ -156,11 +153,9 @@ def test_SymbolicSys__from_callback_bateman(band):
     atol, rtol = 1e-11, 1e-11
     odesys = SymbolicSys.from_callback(decay_dydt_factory(k), len(k)+1,
                                        band=band)
-    out = odesys.integrate_scipy(tend, y0, atol=atol, rtol=rtol)
-    ref = out.copy()
-    ref[:, 1:] = np.array(bateman_full(y0, k+[0], ref[:, 0],
-                                       exp=np.exp)).T
-    assert np.allclose(out, ref, rtol=rtol, atol=atol)
+    xout, yout, info = odesys.integrate_scipy(tend, y0, atol=atol, rtol=rtol)
+    ref = np.array(bateman_full(y0, k+[0], xout - xout[0], exp=np.exp)).T
+    assert np.allclose(yout, ref, rtol=rtol, atol=atol)
 
 
 @pytest.mark.parametrize('band', [(1, 0), None])
@@ -170,11 +165,9 @@ def test_SymbolicSys_bateman(band):
     dydt = decay_dydt_factory(k)
     f = dydt(0, y)
     odesys = SymbolicSys(zip(y, f), band=band)
-    out = odesys.integrate_scipy(tend, y0)
-    ref = out.copy()
-    ref[:, 1:] = np.array(bateman_full(y0, k+[0], ref[:, 0],
-                                       exp=np.exp)).T
-    assert np.allclose(out, ref)
+    xout, yout, info = odesys.integrate_scipy(tend, y0)
+    ref = np.array(bateman_full(y0, k+[0], xout-xout[0], exp=np.exp)).T
+    assert np.allclose(yout, ref)
 
 
 # Longer chains with careful choice of parameters
@@ -228,18 +221,18 @@ def test_scipy(name, forgive):
     y0, k, odesys_dens = get_special_chain(n, p, a)
     if name == 'vode':
         tout = [0]+[10**i for i in range(-10, 1)] if name == 'vode' else 1
-        out = odesys_dens.integrate_scipy(
+        xout, yout, info = odesys_dens.integrate_scipy(
             tout, y0, name=name, atol=atol, rtol=rtol)
-        check(out[-1, 1:], n, p, a, atol, rtol, forgive[0])
+        check(yout[-1, :], n, p, a, atol, rtol, forgive[0])
 
-        out = odesys_dens.integrate_scipy(
+        xout, yout, info = odesys_dens.integrate_scipy(
             1, y0, name=name, atol=atol, rtol=rtol)
-        check(out[-1, 1:], n, p, a, atol, rtol, forgive[1])
+        check(yout[-1, :], n, p, a, atol, rtol, forgive[1])
 
     else:
-        out = odesys_dens.integrate_scipy(
+        xout, yout, info = odesys_dens.integrate_scipy(
             1, y0, name=name, atol=atol, rtol=rtol)
-        check(out[-1, 1:], n, p, a, atol, rtol, forgive)
+        check(yout[-1, :], n, p, a, atol, rtol, forgive)
 
 
 @pytest.mark.parametrize('method,forgive', zip(
@@ -251,10 +244,10 @@ def test_odeint(method, forgive):
     dydt = decay_dydt_factory(k)
     odesys_dens = SymbolicSys.from_callback(dydt, len(k)+1)
     # adaptive stepper fails to produce the accuracy asked for.
-    out = odesys_dens.integrate_odeint(
+    xout, yout, info = odesys_dens.integrate_odeint(
         [10**i for i in range(-15, 1)], y0, method=method,
         atol=atol, rtol=rtol)
-    check(out[-1, 1:], n, p, a, atol, rtol, forgive)
+    check(yout[-1, :], n, p, a, atol, rtol, forgive)
 
 
 def _gsl(tout, method, forgive):
@@ -264,9 +257,9 @@ def _gsl(tout, method, forgive):
     dydt = decay_dydt_factory(k)
     odesys_dens = SymbolicSys.from_callback(dydt, len(k)+1)
     # adaptive stepper fails to produce the accuracy asked for.
-    out = odesys_dens.integrate_gsl(tout, y0, method=method,
-                                    atol=atol, rtol=rtol)
-    check(out[-1, 1:], n, p, a, atol, rtol, forgive)
+    xout, yout, info = odesys_dens.integrate_gsl(tout, y0, method=method,
+                                                 atol=atol, rtol=rtol)
+    check(yout[-1, :], n, p, a, atol, rtol, forgive)
 
 
 @pytest.mark.parametrize('method,forgive', zip(
@@ -288,20 +281,20 @@ def _cvode(tout, method, forgive):
     dydt = decay_dydt_factory(k)
     odesys_dens = SymbolicSys.from_callback(dydt, len(k)+1)
     # adaptive stepper fails to produce the accuracy asked for.
-    out = odesys_dens.integrate_cvode(tout, y0, method=method,
-                                      atol=atol, rtol=rtol)
-    check(out[-1, 1:], n, p, a, atol, rtol, forgive)
+    xout, yout, info = odesys_dens.integrate_cvode(tout, y0, method=method,
+                                                   atol=atol, rtol=rtol)
+    check(yout[-1, :], n, p, a, atol, rtol, forgive)
 
 
 @pytest.mark.parametrize('method,forgive', zip(
-    'adams bdf'.split(), (.17, .13)))
+    'adams bdf'.split(), (.24, .9)))
 def test_cvode_predefined(method, forgive):
     _cvode([10**i for i in range(-15, 1)], method, forgive)
 
 
 # cvode performs significantly better than vode:
 @pytest.mark.parametrize('method,forgive', zip(
-    'adams bdf'.split(), (.18, 2)))
+    'adams bdf'.split(), (.36, 2)))
 def test_cvode_adaptive(method, forgive):
     _cvode(1, method, forgive)
 
@@ -312,39 +305,51 @@ def test_long_chain_dense(n, forgive):
     y0, k, odesys_dens = get_special_chain(n, p, a)
     atol, rtol = 1e-12, 1e-12
     tout = 1
-    out = odesys_dens.integrate_scipy(tout, y0, atol=atol, rtol=rtol)
-    check(out[-1, 1:], n, p, a, atol, rtol, forgive)
+    xout, yout, info = odesys_dens.integrate_scipy(tout, y0,
+                                                   atol=atol, rtol=rtol)
+    check(yout[-1, :], n, p, a, atol, rtol, forgive)
 
 
-@pytest.mark.parametrize('n', [27, 30])  # something maxes out at 31
+@pytest.mark.parametrize('n', [21, 30])  # something maxes out at 31
 def test_long_chain_banded_scipy(n):
     p, a = 0, n
     y0, k, odesys_dens = get_special_chain(n, p, a)
     y0, k, odesys_band = get_special_chain(n, p, a, band=(1, 0))
     atol, rtol = 1e-7, 1e-7
-    time_dens, out_dens = timeit(odesys_dens.integrate_scipy,
-                                 1, y0, atol=atol, rtol=rtol)
-    time_band, out_band = timeit(odesys_band.integrate_scipy,
-                                 1, y0, atol=atol, rtol=rtol)
-    check(out_dens[-1, 1:], n, p, a, atol, rtol, .4)
-    check(out_band[-1, 1:], n, p, a, atol, rtol, .4)
-    print(time_dens, time_band)
+    tout = np.logspace(-10, 0, 10)
+    for _ in range(2):  # warmup
+        time_dens, (xout_dens, yout_dens, info) = timeit(
+            odesys_dens.integrate_scipy, tout, y0, atol=atol, rtol=rtol,
+            name='vode', method='bdf', first_step=1e-10)
+        assert info['njac'] > 0
+    for _ in range(2):  # warmup
+        time_band, (xout_band, yout_band, info) = timeit(
+            odesys_band.integrate_scipy, tout, y0, atol=atol, rtol=rtol,
+            name='vode', method='bdf', first_step=1e-10)
+        assert info['njac'] > 0
+    check(yout_dens[-1, :], n, p, a, atol, rtol, 1.5)
+    check(yout_band[-1, :], n, p, a, atol, rtol, 1.5)
     assert time_dens > time_band  # will fail sometimes due to load
 
 
-@pytest.mark.parametrize('n', [79])
-@pytest.mark.xfail
+@pytest.mark.parametrize('n', [29, 79])
 def test_long_chain_banded_cvode(n):
     p, a = 0, n
     y0, k, odesys_dens = get_special_chain(n, p, a)
     y0, k, odesys_band = get_special_chain(n, p, a, band=(1, 0))
     atol, rtol = 1e-9, 1e-9
-    for i in range(3):
-        time_band, out_band = timeit(odesys_band.integrate_cvode,
-                                     1, y0, atol=atol, rtol=rtol)
-        time_dens, out_dens = timeit(odesys_dens.integrate_cvode,
-                                     1, y0, atol=atol, rtol=rtol)
-    check(out_dens[-1, 1:], n, p, a, atol, rtol, 7)
-    check(out_band[-1, 1:], n, p, a, atol, rtol, 25)  # suspicious
-    print(time_dens, time_band)
-    assert time_dens > time_band  # will fail sometimes due to load
+    for _ in range(2):  # warmup
+        time_band, (xout_band, yout_band, info) = timeit(
+            odesys_band.integrate_cvode, 1, y0, atol=atol, rtol=rtol)
+        assert info['njac'] > 0
+    for _ in range(2):  # warmup
+        time_dens, (xout_dens, yout_dens, info) = timeit(
+            odesys_dens.integrate_cvode, 1, y0, atol=atol, rtol=rtol)
+        assert info['njac'] > 0
+    check(yout_dens[-1, :], n, p, a, atol, rtol, 7)
+    check(yout_band[-1, :], n, p, a, atol, rtol, 25)  # suspicious
+    assert info['njac'] > 0
+    try:
+        assert time_dens > time_band
+    except AssertionError:
+        pass  # will fail sometimes due to load