+++ scripts/run_tests.sh
@@ -1,5 +1,8 @@
-#!/bin/bash
-cd $(dirname $0)/..
-python2.7 -m pytest --ignore build/ --ignore doc/ $@
-python3 -m pytest --ignore build/ --ignore doc/ --doctest-modules --pep8 --flakes $@
+#!/bin/bash -e
+# Usage
+#   $ ./scripts/run_tests.sh
+# or
+#   $ ./scripts/run_tests.sh --cov pycvodes --cov-report html
+python2.7 -m pytest
+python3 -m pytest --doctest-modules --pep8 --flakes $@
 python3 -m doctest README.rst